Move ESIL and analysis APIs to the corresponding places
